# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Meal Planner API",
    "description": "API for managing recipes, meal plans, and shopping lists",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/recipes": {
      "get": {
        "summary": "List all recipes",
        "tags": [
          "Recipes"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number"
                      },
                      "name": {
                        "type": "string"
                      },
                      "description": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "nullable": true,
                            "enum": [
                              null
                            ]
                          }
                        ]
                      },
                      "protein": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "nullable": true,
                            "enum": [
                              null
                            ]
                          }
                        ]
                      },
                      "mealTypes": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "suitableDays": {
                        "type": "string",
                        "enum": [
                          "any",
                          "weekday",
                          "weekend"
                        ]
                      },
                      "freezable": {
                        "type": "boolean"
                      },
                      "lastUsedAt": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "nullable": true,
                            "enum": [
                              null
                            ]
                          }
                        ]
                      },
                      "createdAt": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "nullable": true,
                            "enum": [
                              null
                            ]
                          }
                        ]
                      },
                      "updatedAt": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "nullable": true,
                            "enum": [
                              null
                            ]
                          }
                        ]
                      },
                      "ingredients": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "number"
                            },
                            "name": {
                              "type": "string"
                            },
                            "category": {
                              "anyOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "nullable": true,
                                  "enum": [
                                    null
                                  ]
                                }
                              ]
                            },
                            "units": {
                              "anyOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "nullable": true,
                                  "enum": [
                                    null
                                  ]
                                }
                              ]
                            },
                            "quantity": {
                              "anyOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "nullable": true,
                                  "enum": [
                                    null
                                  ]
                                }
                              ]
                            },
                            "notes": {
                              "anyOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "nullable": true,
                                  "enum": [
                                    null
                                  ]
                                }
                              ]
                            }
                          },
                          "required": [
                            "id",
                            "name",
                            "category",
                            "units",
                            "quantity",
                            "notes"
                          ],
                          "additionalProperties": false
                        }
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "description",
                      "protein",
                      "mealTypes",
                      "suitableDays",
                      "freezable",
                      "lastUsedAt",
                      "createdAt",
                      "updatedAt",
                      "ingredients"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new recipe",
        "tags": [
          "Recipes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "description": {
                    "type": "string"
                  },
                  "protein": {
                    "type": "string"
                  },
                  "mealTypes": {
                    "default": [
                      "dinner"
                    ],
                    "minItems": 1,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "breakfast",
                        "lunch",
                        "dinner"
                      ]
                    }
                  },
                  "suitableDays": {
                    "default": "any",
                    "type": "string",
                    "enum": [
                      "any",
                      "weekday",
                      "weekend"
                    ]
                  },
                  "freezable": {
                    "default": false,
                    "type": "boolean"
                  },
                  "ingredients": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string",
                          "minLength": 1
                        },
                        "quantity": {
                          "type": "string"
                        },
                        "notes": {
                          "type": "string"
                        },
                        "category": {
                          "type": "string"
                        },
                        "units": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "name"
                      ]
                    }
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    },
                    "protein": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    },
                    "mealTypes": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "suitableDays": {
                      "type": "string",
                      "enum": [
                        "any",
                        "weekday",
                        "weekend"
                      ]
                    },
                    "freezable": {
                      "type": "boolean"
                    },
                    "lastUsedAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    },
                    "createdAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    },
                    "updatedAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    },
                    "ingredients": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number"
                          },
                          "name": {
                            "type": "string"
                          },
                          "category": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "nullable": true,
                                "enum": [
                                  null
                                ]
                              }
                            ]
                          },
                          "units": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "nullable": true,
                                "enum": [
                                  null
                                ]
                              }
                            ]
                          },
                          "quantity": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "nullable": true,
                                "enum": [
                                  null
                                ]
                              }
                            ]
                          },
                          "notes": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "nullable": true,
                                "enum": [
                                  null
                                ]
                              }
                            ]
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "category",
                          "units",
                          "quantity",
                          "notes"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "description",
                    "protein",
                    "mealTypes",
                    "suitableDays",
                    "freezable",
                    "lastUsedAt",
                    "createdAt",
                    "updatedAt",
                    "ingredients"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/recipes/import": {
      "post": {
        "summary": "Import recipes from CSV file",
        "tags": [
          "Recipes"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "created": {
                      "type": "number"
                    },
                    "skipped": {
                      "type": "number"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "row": {
                            "type": "number"
                          },
                          "name": {
                            "type": "string"
                          },
                          "error": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "row",
                          "name",
                          "error"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "created",
                    "skipped",
                    "errors"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {}
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}